refactor(auth): relocate NostrIdentity hierarchy into lib/services/auth/ (#4741)
Pure move of the sealed NostrIdentity hierarchy (453 lines, unchanged) plus
its mirrored test into the auth/ collaborator folder created by #5729, so the
signer-core extraction lands next to it. 26 import rewrites, no behavior
change. Ratchet-neutral: the untested-services floor matches basenames
recursively and nostr_identity_test.dart keeps its basename.
