+++ mobile/lib/services/auth/nostr_identity.dart
