+++ mobile/test/helpers/real_integration_test_helper.dart
@@ -5,7 +5,7 @@ import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:nostr_client/nostr_client.dart';
 import 'package:nostr_key_manager/nostr_key_manager.dart';
-import 'package:openvine/services/nostr_identity.dart';
+import 'package:openvine/services/auth/nostr_identity.dart';
 import 'package:openvine/services/nostr_service_factory.dart';
 
 /// Setup real integration test environment with minimal mocking
