+++ mobile/test/services/auth_service_bunker_connect_test.dart
@@ -18,8 +18,8 @@ import 'package:mocktail/mocktail.dart';
 import 'package:nostr_sdk/nostr_sdk.dart'
     show NostrRemoteSigner, NostrRemoteSignerInfo;
 import 'package:openvine/models/known_account.dart';
+import 'package:openvine/services/auth/nostr_identity.dart';
 import 'package:openvine/services/auth_service.dart';
-import 'package:openvine/services/nostr_identity.dart';
 import 'package:openvine/services/user_data_cleanup_service.dart';
 import 'package:shared_preferences/shared_preferences.dart';
