+++ mobile/test/services/nostr_service_factory_test.dart
@@ -4,7 +4,7 @@ import 'package:nostr_client/nostr_client.dart';
 import 'package:nostr_key_manager/nostr_key_manager.dart';
 import 'package:nostr_sdk/nostr_sdk.dart';
 import 'package:openvine/models/environment_config.dart';
-import 'package:openvine/services/nostr_identity.dart';
+import 'package:openvine/services/auth/nostr_identity.dart';
 import 'package:openvine/services/nostr_service_factory.dart';
 
 class _MockSecureKeyContainer extends Mock implements SecureKeyContainer {}
